backend/pyproject.toml
[project]
name = "grocery-genie-backend"
version = "0.1.0"
description = "Grocery Genie API — receipt scanning and budget tracking"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "celery[redis]>=5.4.0",
    "redis>=5.1.0",
    "litellm>=1.50.0",
    "pillow>=10.4.0",
    "opencv-python-headless>=4.10.0",
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.30.0",
    "alembic>=1.13.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "aiosqlite>=0.20.0",
    "ruff>=0.7.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.lint.per-file-ignores]
"app/api/routes/*.py" = ["B008"]  # Depends() in function defaults is standard FastAPI
"app/core/security.py" = ["B008"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
